+++ src/main/java/cart/application/CouponService.java
@@ -1,8 +1,8 @@
 package cart.application;
 
 import cart.dao.CouponDao;
-import cart.domain.Coupon;
-import cart.domain.Discount;
+import cart.domain.coupon.Coupon;
+import cart.domain.coupon.Discount;
 import cart.dto.coupon.CouponRequest;
 import org.springframework.stereotype.Service;
