+++ src/main/java/cart/dao/MemberDao.java
@@ -1,13 +1,12 @@
 package cart.dao;
 
-import cart.domain.Member;
+import cart.domain.member.Member;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
 
 @Repository
