+++ src/main/java/cart/dao/OrderCouponDao.java
@@ -1,7 +1,7 @@
 package cart.dao;
 
-import cart.domain.Coupon;
-import cart.domain.Discount;
+import cart.domain.coupon.Coupon;
+import cart.domain.coupon.Discount;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
