+++ src/main/java/cart/domain/cartItem/CartItem.java
@@ -1,26 +1,27 @@
-package cart.domain;
+package cart.domain.cartItem;
 
+import cart.domain.member.Member;
+import cart.domain.product.Product;
 import cart.exception.CartItemException;
 
 import java.util.Objects;
 
 public class CartItem {
-    private Long id;
-    private int quantity;
+
+    private final Long id;
+    private final int quantity;
     private final Product product;
     private final Member member;
-
-    public CartItem(Member member, Product product) {
-        this.quantity = 1;
-        this.member = member;
-        this.product = product;
-    }
 
     public CartItem(Long id, int quantity, Product product, Member member) {
         this.id = id;
         this.quantity = quantity;
         this.product = product;
         this.member = member;
+    }
+
+    public CartItem(Member member, Product product) {
+        this(null, 1, product, member);
     }
 
     public Long getId() {
@@ -45,7 +46,7 @@ public class CartItem {
         }
     }
 
-    public void changeQuantity(int quantity) {
-        this.quantity = quantity;
+    public CartItem changeQuantity(int quantity) {
+        return new CartItem(id, quantity, product, member);
     }
 }
