+++ src/main/java/cart/domain/coupon/Discount.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.coupon;
 
 public class Discount {
