+++ src/main/java/cart/domain/order/Order.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.order;
 
 import java.util.List;
