+++ src/main/java/cart/domain/product/Product.java
@@ -1,22 +1,21 @@
-package cart.domain;
+package cart.domain.product;
 
 public class Product {
-    private Long id;
-    private String name;
-    private int price;
-    private String imageUrl;
 
-    public Product(String name, int price, String imageUrl) {
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
+    private final Long id;
+    private final String name;
+    private final int price;
+    private final String imageUrl;
 
     public Product(Long id, String name, int price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
+    }
+
+    public Product(String name, int price, String imageUrl) {
+        this(null, name, price, imageUrl);
     }
 
     public Long getId() {
