+++ src/main/java/cart/dto/coupon/DiscountResponse.java
@@ -1,6 +1,6 @@
 package cart.dto.coupon;
 
-import cart.domain.Discount;
+import cart.domain.coupon.Discount;
 
 public class DiscountResponse {
