+++ src/main/java/cart/repository/OrderRepository.java
@@ -3,10 +3,10 @@ package cart.repository;
 import cart.dao.OrderCouponDao;
 import cart.dao.OrderDao;
 import cart.dao.OrderItemDao;
-import cart.domain.Coupon;
-import cart.domain.Order;
-import cart.domain.OrderItem;
-import cart.domain.Product;
+import cart.domain.coupon.Coupon;
+import cart.domain.order.Order;
+import cart.domain.order.OrderItem;
+import cart.domain.product.Product;
 import cart.entity.OrderItemEntity;
 import cart.entity.ProductEntity;
 import org.springframework.stereotype.Repository;
