+++ src/main/java/cart/ui/CartItemApiController.java
@@ -1,7 +1,7 @@
 package cart.ui;
 
 import cart.application.CartItemService;
-import cart.domain.Member;
+import cart.domain.member.Member;
 import cart.dto.cartItem.CartItemQuantityUpdateRequest;
 import cart.dto.cartItem.CartItemRequest;
 import cart.dto.cartItem.CartItemResponse;
