+++ src/main/java/cart/ui/OrderController.java
@@ -1,8 +1,8 @@
 package cart.ui;
 
 import cart.application.OrderService;
-import cart.domain.Member;
-import cart.domain.Order;
+import cart.domain.member.Member;
+import cart.domain.order.Order;
 import cart.dto.order.OrderRequest;
 import cart.dto.order.OrderResponse;
 import org.springframework.http.ResponseEntity;
